Complete CSS foundations
I completed part 6 - cascade fix. I now have a pretty good understanding
of how cascades influence style - both in the position in the code
(where selectors lie from top to bottom) and also in the specificity
of any given selector.

* No level of specificity
. Lowest level of specificity
\# Higher level of specificty

If an element has multiple selectors, whatever rules share the most
selectors in common will be applied. If an equal number of selectors is
present, the rules at the bottom will apply.