COMPLETE FIRST TWO FLEX EXERCISES
I need to review commit messages - are the titles supposed to be all
caps?
I completed exercise 1 and 2 of the flex exercises. I should probably
re-do the second one to ensure that I fully understand the importance of
setting margin and padding on both a larger and more granular scale. I
also need to work on setting rules higher up in the cascade so that I
don't have to repeat the same rule multiple times.