Complete Flexercise 3
I completed exercise 3, though when I compared my code to the solution,
there were some discrepancies. I need to remember the concept of
grouping different selectors that have the same properties and values.
I am committed to not getting stuck in the weeds here - although the CSS
learnin is kinda tough, I'm going to push through to the end of the
curriculum.